move components with no supplementary files out of their directories
