+++ frontend/src/components/profile/ProfileSectionFooter.tsx
