+++ frontend/src/components/profile/profileSettings/ShipmentAddressSettings.tsx
@@ -1,6 +1,6 @@
 import { ChangeEvent, useState } from 'react';
-import PageSection from '../../../pageSection/PageSection';
-import ProfileSectionFooter from '../../profileSectionFooter/ProfileSectionFooter';
+import PageSection from '../../PageSection';
+import ProfileSectionFooter from '../ProfileSectionFooter';
 import { Button, TextField } from '@mui/material';
 
 function ShipmentAddressSettings(): JSX.Element {
